+++ RELEASE-NOTES.md
@@ -84,16 +84,17 @@ class Args {
 - #196 API Change: `Option::type()` and `Parameters::type()` now return empty array by default (was `{String.class}`).
 - #196 New feature: Infer type from collections and maps when `type` annotation not specified. Thanks to [ddimtirov](https://github.com/ddimtirov) for the suggestion.
 - #197 New feature: Use `type` attribute to determine conversion target type instead of field type. This allows fields to be declared as interfaces or abstract types (or arrays/collections/maps of these) and via the `type` attribute picocli will be able to convert String arguments to concrete implementation objects.
+- #187 New feature: Added methods to programmatically get and set the command name
 - #192 Bugfix: Default arity should be 1, not *, for array and collection options. Thanks to [RobertZenz](https://github.com/RobertZenz).
 - #193 Bugfix: Splitting an argument should not cause max arity to be exceeded.
 - #191 Bugfix: Arity should not limit the total number of values put in an array or collection. Thanks to [RobertZenz](https://github.com/RobertZenz).
 - #186 Bugfix: Confusing usage message for collection options. Thanks to [AlexFalappa](https://github.com/AlexFalappa).
 - #181 Bugfix: Incorrect help message was displayed for short options with paramLabel when arity > 1
+- #184 Bugfix/Enhancement: Programmatically setting the separator is now reflected in the usage help message. Thanks to [defnull](https://github.com/defnull).
 - #194 Enhancement: Usage help should show split regex for option/parameters
 - #198 Enhancement: Usage help parameter list details should indicate arity for positional parameters
 - #195 Enhancement: Usage help should show Map types if paramLabel not specified
 - #185 Enhancement: Exception message text for missing options should not use field names but be more descriptive and consistent with usage help. Thanks to [AlexFalappa](https://github.com/AlexFalappa).
-- #184 Doc: Improved CommandLine.setSeparator javadoc to clarify that this affects parsing only and link to the `@Command` `separator` annotation attribute. Thanks to [defnull](https://github.com/defnull).
 
 ## 1.0.1 - Bugfix release.